+++ main/java/com/springapp/mvc/pojo/exam/Question.java
@@ -6,11 +6,8 @@ import org.hibernate.annotations.Cascade;
 
 import javax.persistence.*;
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.Date;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 /**
  * Created by Phuthikorn_T on 6/30/2015.
@@ -30,7 +27,7 @@ public class Question implements Serializable, Cloneable {
     private String description;
 
     @Column(name = "QUESTION_SCORE")
-    private BigDecimal score;
+    private Float score;
 
     @Column(name = "QUESTION_CREATE_DATE")
     private Date createDate;
@@ -169,11 +166,11 @@ public class Question implements Serializable, Cloneable {
         this.description = description;
     }
 
-    public BigDecimal getScore() {
+    public Float getScore() {
         return score;
     }
 
-    public void setScore(BigDecimal score) {
+    public void setScore(Float score) {
         this.score = score;
     }
